added folder interfaces, moved InStudentRepository there
