+++ src/main/java/service/util/Validation.java
@@ -16,14 +16,14 @@ public class Validation {
          */
         List<String> errors = new ArrayList<>();
 
-        if (testResultRequest.getTestId() < 0) {
-            errors.add("entity.Test ID must be greater than 0");
+        if (testResultRequest.getTestId() <= 0) {
+            errors.add("Entity Test ID must be greater than 0");
         }
 
-        if (testResultRequest.getStudentId() < 0) {
+        if (testResultRequest.getStudentId() <= 0) {
             errors.add("Student ID must be greater than 0");
         }
-        if (testResultRequest.getResult() < 0) {
+        if (testResultRequest.getResult() <= 0) {
             errors.add("Result must be greater than 0");
         }
